Created APIs related to Vending Machine transaction workflow.

+++ backend/src/main/java/com/backend/model/BalanceModel.java
@@ -27,6 +27,6 @@ public class BalanceModel {
     @OneToOne
     @JoinColumn(name="people_id")
     PeopleModel peopleModel;
-    Integer currentBalance;
+    Float currentBalance;
 
 }
